backend/requirements.txt
fastapi
uvicorn
pillow
numpy
scipy
opencv-python
python-multipart